Moves assert_lua! family into helper module
