+++ src/lua/line.rs
@@ -70,7 +70,6 @@ impl UserData for Line {
 }
 
 #[cfg(test)]
-#[allow(unused_macros)]
 mod test_lua_line {
     use rlua::Lua;
 
@@ -88,41 +87,6 @@ mod test_lua_line {
                 .unwrap();
         });
         state
-    }
-
-    macro_rules! test_lua {
-        () => {
-            let state = get_lua();
-
-            macro_rules! assert_lua {
-                ($return_type:ty, $lua_code:literal, $expect:expr) => {
-                    assert_eq!(
-                        state.context(|ctx| -> $return_type {
-                            ctx.load(concat!("return ", $lua_code)).call(()).unwrap()
-                        }),
-                        $expect
-                    );
-                };
-            }
-
-            macro_rules! assert_lua_bool {
-                ($lua_code:literal, $expect:expr) => {
-                    assert_lua!(bool, $lua_code, $expect)
-                };
-            }
-
-            macro_rules! assert_lua_string {
-                ($lua_code:literal, $expect:expr) => {
-                    assert_lua!(String, $lua_code, $expect)
-                };
-            }
-
-            macro_rules! global {
-                ($key:literal) => {
-                    state.context(|ctx| ctx.globals().get($key).unwrap())
-                };
-            }
-        };
     }
 
     #[test]
